+++ src/types/Feed.ts
@@ -1,3 +1,5 @@
+import {ChainsIds} from './ChainsIds';
+
 export interface Feed {
   symbol?: string;
   discrepancy: number;
@@ -43,9 +45,12 @@ export interface PairWithFreshness extends Pair {
 }
 
 export interface OnChainCall {
+  chainId?: ChainsIds; // default ETH
   address: string;
   method: string;
-  inputs: string[];
-  outputs: string[];
+  inputs: string[]; // array of types
+  outputs: string[]; // array of types
   args: string[];
+  returnIndex?: number; // id/index of on-chain returned data that should be used as returned value
+  decimals?: number; // decimals of returned number, if undefined will be returned as string
 }